5.2.1 refactor by D7 criteria
