+++ src/components/comments.js
@@ -1,4 +1,4 @@
-import {EMOJIS} from '../const.js';
+import {EMOJIS} from '../utils/const.js';
 import AbstractComponent from './abstract-component.js';
 
 export default class Comments extends AbstractComponent {
@@ -6,6 +6,30 @@ export default class Comments extends AbstractComponent {
     super();
 
     this._comments = comments;
+  }
+
+  getTemplate() {
+    const commentCount = this._comments.length;
+    const commentMarkup = this._comments.map((it) => this._createCommentMarkup(it)).join(`\n`);
+    const emojiListMarkup = EMOJIS.map((it) => this._createEmojiListMarkup(it)).join(`\n`);
+
+    return (
+      `<section class="film-details__comments-wrap">
+        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentCount}</span></h3>
+        <ul class="film-details__comments-list">
+          ${commentMarkup}
+        </ul>
+        <div class="film-details__new-comment">
+          <div for="add-emoji" class="film-details__add-emoji-label"></div>
+          <label class="film-details__comment-label">
+            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+          </label>
+          <div class="film-details__emoji-list">
+            ${emojiListMarkup}
+          </div>
+        </div>
+      </section>`
+    );
   }
 
   _createCommentMarkup(comment) {
@@ -35,30 +59,6 @@ export default class Comments extends AbstractComponent {
       <label class="film-details__emoji-label" for="emoji-${emoji}">
         <img src="./images/emoji/${emoji}.png" width="30" height="30" alt="emoji">
       </label>`
-    );
-  }
-
-  getTemplate() {
-    const commentCount = this._comments.length;
-    const commentMarkup = this._comments.map((it) => this._createCommentMarkup(it)).join(`\n`);
-    const emojiListMarkup = EMOJIS.map((it) => this._createEmojiListMarkup(it)).join(`\n`);
-
-    return (
-      `<section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentCount}</span></h3>
-        <ul class="film-details__comments-list">
-          ${commentMarkup}
-        </ul>
-        <div class="film-details__new-comment">
-          <div for="add-emoji" class="film-details__add-emoji-label"></div>
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-          <div class="film-details__emoji-list">
-            ${emojiListMarkup}
-          </div>
-        </div>
-      </section>`
     );
   }
 }
