+++ src/components/navigation.js
@@ -14,16 +14,6 @@ export default class Navigation extends AbstractComponent {
     this._quantity = quantity;
   }
 
-  _createNavigationMarkup(navItem) {
-    const [id, name] = navItem;
-    const count = this._quantity[id] ? this._quantity[id] : 0;
-    return (
-      `<a href="#${id}" class="main-navigation__item ${id === `all` ? `main-navigation__item--active` : ``}">
-        ${name} ${id === `all` ? `` : `<span class="main-navigation__item-count">${count}</span>`}
-      </a>\n`
-    );
-  }
-
   getTemplate() {
     const navigationMarkup = [...NAV_ITEMS].map((name) => this._createNavigationMarkup(name)).join(`\n`);
 
@@ -34,6 +24,16 @@ export default class Navigation extends AbstractComponent {
         </div>
         <a href="#stats" class="main-navigation__additional">Stats</a>
       </nav>`
+    );
+  }
+
+  _createNavigationMarkup(navItem) {
+    const [id, name] = navItem;
+    const count = this._quantity[id] ? this._quantity[id] : 0;
+    return (
+      `<a href="#${id}" class="main-navigation__item ${id === `all` ? `main-navigation__item--active` : ``}">
+        ${name} ${id === `all` ? `` : `<span class="main-navigation__item-count">${count}</span>`}
+      </a>\n`
     );
   }
 }
