+++ src/components/show-more-button.js
@@ -1,13 +1,13 @@
 import AbstractComponent from './abstract-component.js';
 
 export default class ShowMoreButton extends AbstractComponent {
-  setClickHandler(handler) {
-    this.getElement().addEventListener(`click`, handler);
-  }
-
   getTemplate() {
     return (
       `<button class="films-list__show-more">Show more</button>`
     );
+  }
+
+  setClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
   }
 }
