+++ src/controllers/page.js
@@ -28,6 +28,47 @@ export default class PageController {
     this._openedFilmDetails = false;
   }
 
+  render(movies) {
+    const isDataBaseEmpty = movies.length === 0 ? true : false;
+    if (isDataBaseEmpty) {
+      render(this._container, this._noDataComponent);
+      return;
+    }
+
+    render(this._container, this._filmContainerComponent);
+
+    const filmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list`);
+    const filmListContainerElement = filmListElement.querySelector(`.films-list__container`);
+
+    let showingFilmCount = FILM_COUNT_ON_START;
+    movies.slice(0, showingFilmCount).forEach((movie) => {
+      this._renderFilm(filmListContainerElement, movie);
+    });
+
+    render(filmListElement, this._showMoreButtonComponent);
+
+    this._showMoreButtonComponent.setClickHandler((evt) => {
+      evt.preventDefault();
+      const prevFilmCount = showingFilmCount;
+      showingFilmCount += FILM_COUNT_BY_BUTTON;
+
+      movies.slice(prevFilmCount, showingFilmCount).forEach((movie) => {
+        this._renderFilm(filmListContainerElement, movie);
+      });
+
+      if (showingFilmCount >= movies.length) {
+        remove(this._showMoreButtonComponent);
+      }
+    });
+
+    const topFilms = getSortedFilms(movies, SortType.RATING);
+    const commentedFilms = getSortedFilms(movies, SortType.COMMENTS);
+    const topFilmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list--top .films-list__container`);
+    const commentedFilmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list--commented .films-list__container`);
+    this._repeatRender(topFilmListElement, topFilms, TOP_FILM_COUNT);
+    this._repeatRender(commentedFilmListElement, commentedFilms, COMMENTED_FILM_COUNT);
+  }
+
   _renderFilm(place, movie) {
 
     const openFilmDetails = (evt) => {
@@ -77,46 +118,5 @@ export default class PageController {
     for (let i = 0; i < iterations; i++) {
       this._renderFilm(container, movies[i]);
     }
-  }
-
-  render(movies) {
-    const isDataBaseEmpty = movies.length === 0 ? true : false;
-    if (isDataBaseEmpty) {
-      render(this._container, this._noDataComponent);
-      return;
-    }
-
-    render(this._container, this._filmContainerComponent);
-
-    const filmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list`);
-    const filmListContainerElement = filmListElement.querySelector(`.films-list__container`);
-
-    let showingFilmCount = FILM_COUNT_ON_START;
-    movies.slice(0, showingFilmCount).forEach((movie) => {
-      this._renderFilm(filmListContainerElement, movie);
-    });
-
-    render(filmListElement, this._showMoreButtonComponent);
-
-    this._showMoreButtonComponent.setClickHandler((evt) => {
-      evt.preventDefault();
-      const prevFilmCount = showingFilmCount;
-      showingFilmCount += FILM_COUNT_BY_BUTTON;
-
-      movies.slice(prevFilmCount, showingFilmCount).forEach((movie) => {
-        this._renderFilm(filmListContainerElement, movie);
-      });
-
-      if (showingFilmCount >= movies.length) {
-        remove(this._showMoreButtonComponent);
-      }
-    });
-
-    const topFilms = getSortedFilms(movies, SortType.RATING);
-    const commentedFilms = getSortedFilms(movies, SortType.COMMENTS);
-    const topFilmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list--top .films-list__container`);
-    const commentedFilmListElement = this._filmContainerComponent.getElement().querySelector(`.films-list--commented .films-list__container`);
-    this._repeatRender(topFilmListElement, topFilms, TOP_FILM_COUNT);
-    this._repeatRender(commentedFilmListElement, commentedFilms, COMMENTED_FILM_COUNT);
   }
 }
